+++ CHANGES.rst
@@ -6,6 +6,9 @@ New Features
 
 - Profile viewers now support plotting with profiles "as steps". [#1595]
 
+- Line flux in the Line Analysis plugin are reported in W/m2 if Spectral Flux is given
+  in Jy [#1564]
+
 Cubeviz
 ^^^^^^^